fix: task

+++ src/synthesis/ffmpeg.ts
@@ -1,17 +1,21 @@
+import { wavsDir } from "../config.ts";
 import { exec } from "../deps.ts";
 
 export const concatAudios = async (
   paths: string[],
   outPath: string
 ): Promise<void> => {
-  const inputTxtPath = `data/_wavs/input.txt`;
+  const inputTxtPath = `${wavsDir}/input.txt`;
   const inputs = paths
     .map((path) => `file '${path.split("/").at(-1)}'`)
     .join("\n");
   await Deno.writeTextFile(inputTxtPath, inputs);
   const cmd = `ffmpeg -y -f concat -i ${inputTxtPath} ${outPath}`;
   console.log(`[ffmpeg] ${cmd}`);
-  await exec(cmd);
-
+  const res = await exec(cmd);
+  if (!res.status.success) {
+    console.log(inputTxtPath);
+    throw res;
+  }
   await Deno.remove(inputTxtPath);
 };
